requirements.txt
boto3==1.20.23
pino==0.6.0
python-dotenv==0.19.2
